requirements.txt
llama-index 
deeplake 
openai 
python-dotenv 
llama-hub
streamlit